Move gate class to a separate file
