+++ src/qutip_qip/operations/__init__.py
@@ -2,3 +2,4 @@
 Operations on quantum circuits.
 """
 from .gates import *
+from .gateclass import *
