+++ src/qutip_qip/operations/gates.py
@@ -14,10 +14,7 @@ import scipy.sparse as sp
 import qutip
 from qutip import Qobj, identity, qeye, sigmax, sigmay, sigmaz, tensor, fock_dm
 
-
 __all__ = [
-    "Gate",
-    "GATE_CLASS_MAP",
     "rx",
     "ry",
     "rz",
@@ -59,704 +56,6 @@ __all__ = [
     "expand_operator",
 ]
 
-
-class Gate:
-    """
-    Representation of a quantum gate, with its required parameters, and target
-    and control qubits.
-    Base class for a quantum gate,
-    concrete gate classes need to be defined as subclasses.
-
-    Parameters
-    ----------
-    targets : list or int
-        The target qubits fo the gate.
-    controls : list or int
-        The controlling qubits of the gate.
-    arg_value : object
-        Argument value of the gate. It will be saved as an attributes and
-        can be accessed when generating the `:obj:qutip.Qobj`.
-    classical_controls : int or list of int, optional
-        Indices of classical bits to control the unitary operator.
-    control_value : int or (int, int), optional
-        The decimal value of controlling bits for executing
-        the unitary operator on the target qubits.
-        If the gate is only quantum controlled or classically controlled,
-        it could be given as an integer,
-        If both are used, it should be a tuple where the first one is quantum
-        and the second one the classical.
-        E.g. if the gate should be executed when the zero-th bit is 1,
-        ``controll_value=1``;
-        If the gate should be executed when the first two bits are 0 and 1,
-        ``controll_value=3``.
-    arg_label : string
-        Label for the argument, it will be shown in the circuit plot,
-        representing the argument value provided to the gate, e.g,
-        if ``arg_label="\phi" the latex name for the gate in the circuit plot
-        will be ``$U(\phi)$``.
-    name : string, optional
-        The name of the gate. This is kept for backward compatibility
-        to identify different gates.
-        In most cases it is identical to the class name,
-        but that is not guaranteed.
-        It is recommended to use ``isinstance``
-        or ``issubclass`` to identify a gate rather than
-        comparing the name string.
-    """
-
-    def __init__(
-        self,
-        name=None,
-        targets=None,
-        controls=None,
-        arg_value=None,
-        control_value=None,
-        classical_controls=None,
-        arg_label=None,
-        **kwargs,
-    ):
-        """
-        Create a gate with specified parameters.
-        """
-
-        self.name = name
-        self.targets = None
-        self.controls = None
-        self.classical_controls = None
-        self.control_value = None
-
-        if not isinstance(targets, Iterable) and targets is not None:
-            self.targets = [targets]
-        else:
-            self.targets = targets
-
-        if not isinstance(controls, Iterable) and controls is not None:
-            self.controls = [controls]
-        else:
-            self.controls = controls
-
-        if (
-            not isinstance(classical_controls, Iterable)
-            and classical_controls is not None
-        ):
-            self.classical_controls = [classical_controls]
-        else:
-            self.classical_controls = classical_controls
-
-        self.control_value = control_value
-        self.arg_value = arg_value
-        self.arg_label = arg_label
-        self.latex_str = r"U"
-
-        for ind_list in [self.targets, self.controls, self.classical_controls]:
-            if ind_list is None:
-                continue
-            all_integer = all(
-                [isinstance(ind, numbers.Integral) for ind in ind_list]
-            )
-            if not all_integer:
-                raise ValueError("Index of a qubit must be an integer")
-
-    def get_all_qubits(self):
-        """
-        Return a list of all qubits that the gate operator
-        acts on.
-        The list concatenates the two lists representing
-        the controls and the targets qubits while retains the order.
-
-        Returns
-        -------
-        targets_list : list of int
-            A list of all qubits, including controls and targets.
-        """
-        if self.controls is not None:
-            return self.controls + self.targets
-        if self.targets is not None:
-            return self.targets
-        else:
-            # Special case: the global phase gate
-            return []
-
-    def __str__(self):
-        str_name = (
-            "Gate(%s, targets=%s, controls=%s,"
-            " classical controls=%s, control_value=%s)"
-        ) % (
-            self.name,
-            self.targets,
-            self.controls,
-            self.classical_controls,
-            self.control_value,
-        )
-        return str_name
-
-    def __repr__(self):
-        return str(self)
-
-    def _repr_latex_(self):
-        return str(self)
-
-    def _to_qasm(self, qasm_out):
-        """
-        Pipe output of gate signature and application to QasmOutput object.
-
-        Parameters
-        ----------
-        qasm_out: QasmOutput
-            object to store QASM output.
-        """
-
-        qasm_gate = qasm_out.qasm_name(self.name)
-
-        if not qasm_gate:
-            error_str = "{} gate's qasm defn is not specified".format(
-                self.name
-            )
-            raise NotImplementedError(error_str)
-
-        if self.classical_controls:
-            err_msg = "Exporting controlled gates is not implemented yet."
-            raise NotImplementedError(err_msg)
-        else:
-            qasm_out.output(
-                qasm_out._qasm_str(
-                    qasm_gate, self.controls, self.targets, self.arg_value
-                )
-            )
-
-    def get_compact_qobj(self):
-        """
-        Get the compact :class:`qutip.Qobj` representation of the gate
-        operator, ignoring the controls and targets.
-        In the unitary representation,
-        it always assumes that the first few qubits are controls,
-        then targets.
-
-        Returns
-        -------
-        qobj : :obj:`qutip.Qobj`
-            The compact gate operator as a unitary matrix.
-        """
-        # TODO This will be moved to each sub-class of Gate
-        if self.name == "RX":
-            qobj = rx(self.arg_value)
-        elif self.name == "RY":
-            qobj = ry(self.arg_value)
-        elif self.name == "RZ":
-            qobj = rz(self.arg_value)
-        elif self.name == "X":
-            qobj = x_gate()
-        elif self.name == "Y":
-            qobj = y_gate()
-        elif self.name == "CY":
-            qobj = cy_gate()
-        elif self.name == "Z":
-            qobj = z_gate()
-        elif self.name == "CZ":
-            qobj = cz_gate()
-        elif self.name == "T":
-            qobj = t_gate()
-        elif self.name == "CT":
-            qobj = ct_gate()
-        elif self.name == "S":
-            qobj = s_gate()
-        elif self.name == "CS":
-            qobj = cs_gate()
-        elif self.name == "SQRTNOT":
-            qobj = sqrtnot()
-        elif self.name == "SNOT":
-            qobj = snot()
-        elif self.name == "PHASEGATE":
-            qobj = phasegate(self.arg_value)
-        elif self.name == "QASMU":
-            qobj = qasmu_gate(self.arg_value)
-        elif self.name == "CRX":
-            qobj = controlled_gate(rx(self.arg_value))
-        elif self.name == "CRY":
-            qobj = controlled_gate(ry(self.arg_value))
-        elif self.name == "CRZ":
-            qobj = controlled_gate(rz(self.arg_value))
-        elif self.name == "CPHASE":
-            qobj = cphase(self.arg_value)
-        elif self.name == "CNOT":
-            qobj = cnot()
-        elif self.name == "CSIGN":
-            qobj = csign()
-        elif self.name == "BERKELEY":
-            qobj = berkeley()
-        elif self.name == "SWAPalpha":
-            qobj = swapalpha(self.arg_value)
-        elif self.name == "SWAP":
-            qobj = swap()
-        elif self.name == "ISWAP":
-            qobj = iswap()
-        elif self.name == "SQRTSWAP":
-            qobj = sqrtswap()
-        elif self.name == "SQRTISWAP":
-            qobj = sqrtiswap()
-        elif self.name == "FREDKIN":
-            qobj = fredkin()
-        elif self.name == "TOFFOLI":
-            qobj = toffoli()
-        elif self.name == "IDLE":
-            qobj = qeye(2)
-        elif self.name == "GLOBALPHASE":
-            raise NotImplementedError(
-                "Globalphase gate has no compack qobj representation."
-            )
-        else:
-            raise NotImplementedError(f"{self.name} is an unknown gate.")
-        return qobj
-
-    def get_qobj(self, num_qubits=None, dims=None):
-        """
-        Get the :class:`qutip.Qobj` representation of the gate operator.
-        The operator is expanded to the full Herbert space according to
-        the controls and targets qubits defined for the gate.
-
-        Parameters
-        ----------
-        num_qubits : int, optional
-            The number of qubits.
-            If not given, use the minimal number of qubits required
-            by the target and control qubits.
-        dims : list, optional
-            A list representing the dimensions of each quantum system.
-            If not given, it is assumed to be an all-qubit system.
-
-        Returns
-        -------
-        qobj : :obj:`qutip.Qobj`
-            The compact gate operator as a unitary matrix.
-        """
-        if self.name == "GLOBALPHASE":
-            if num_qubits is not None:
-                return globalphase(self.arg_value, num_qubits)
-            else:
-                raise ValueError(
-                    "The number of qubits must be provided for "
-                    "global phase gates."
-                )
-
-        all_targets = self.get_all_qubits()
-        if num_qubits is None:
-            num_qubits = max(all_targets) + 1
-        return expand_operator(
-            self.get_compact_qobj(),
-            N=num_qubits,
-            targets=all_targets,
-            dims=dims,
-        )
-
-
-class SingleQubitGate(Gate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        if self.targets is None or len(self.targets) != 1:
-            raise ValueError(
-                f"Gate {self.__class__.__name__} requires one target"
-            )
-        if self.controls:
-            raise ValueError(
-                f"Gate {self.__class__.__name__} cannot have a control"
-            )
-
-
-class X(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "X"
-        self.latex_str = r"X"
-
-    def get_compact_qobj(self):
-        return qutip.sigmax()
-
-
-class Y(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "Y"
-        self.latex_str = r"Y"
-
-    def get_compact_qobj(self):
-        return qutip.sigmay()
-
-
-class Z(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "Z"
-        self.latex_str = r"Z"
-
-    def get_compact_qobj(self):
-        return qutip.sigmaz()
-
-
-class RX(SingleQubitGate):
-    def __init__(self, targets, arg_value, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "RX"
-        self.latex_str = r"R_x"
-
-    def get_compact_qobj(self):
-        return rx(self.arg_value)
-
-
-class RY(SingleQubitGate):
-    def __init__(self, targets, arg_value, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "RY"
-        self.latex_str = r"R_y"
-
-    def get_compact_qobj(self):
-        return ry(self.arg_value)
-
-
-class RZ(SingleQubitGate):
-    def __init__(self, targets, arg_value, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "RZ"
-        self.latex_str = r"R_z"
-
-    def get_compact_qobj(self):
-        return rz(self.arg_value)
-
-
-class H(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "SNOT"
-        self.latex_str = r"{\rm H}"
-
-    def get_compact_qobj(self):
-        return snot()
-
-
-SNOT = H
-
-
-class SQRTNOT(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "SQRTNOT"
-
-    def get_compact_qobj(self):
-        return sqrtnot()
-        self.latex_str = r"\sqrt{\rm NOT}"
-
-
-class S(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "S"
-        self.latex_str = r"{\rm S}"
-
-    def get_compact_qobj(self):
-        return s_gate()
-
-
-class T(SingleQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "T"
-        self.latex_str = r"{\rm T}"
-
-    def get_compact_qobj(self):
-        return t_gate()
-
-
-class QASMU(SingleQubitGate):
-    def __init__(self, targets, arg_value=None, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "QASMU"
-        self.latex_str = r"{\rm QASM-U}"
-
-    def get_compact_qobj(self):
-        return qasmu_gate(self.arg_value)
-
-
-class TwoQubitGate(Gate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        if len(self.get_all_qubits()) != 2:
-            raise ValueError(
-                f"Gate {self.__class__.__name__} requires two targets"
-            )
-
-
-class SWAP(TwoQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "SWAP"
-        self.latex_str = r"{\rm SWAP}"
-
-    def get_compact_qobj(self):
-        return swap()
-
-
-class ISWAP(TwoQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "ISWAP"
-        self.latex_str = r"{i}{\rm SWAP}"
-
-    def get_compact_qobj(self):
-        return iswap()
-
-
-class CNOT(TwoQubitGate):
-    def __init__(self, targets, controls, **kwargs):
-        super().__init__(targets=targets, controls=controls, **kwargs)
-        self.name = "CNOT"
-        self.latex_str = r"{\rm CNOT}"
-
-    def get_compact_qobj(self):
-        return cnot()
-
-
-class SQRTSWAP(TwoQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "SQRTSWAP"
-        self.latex_str = r"\sqrt{\rm SWAP}"
-
-    def get_compact_qobj(self):
-        return sqrtswap()
-
-
-class SQRTISWAP(TwoQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "SQRTISWAP"
-        self.latex_str = r"\sqrt{{i}\rm SWAP}"
-
-    def get_compact_qobj(self):
-        return sqrtiswap()
-
-
-class BERKELEY(TwoQubitGate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "BERKELEY"
-        self.latex_str = r"{\rm BERKELEY}"
-
-    def get_compact_qobj(self):
-        return berkeley()
-
-
-class SWAPALPHA(TwoQubitGate):
-    def __init__(self, targets, arg_value, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "SWAPALPHA"
-        self.latex_str = r"{\rm SWAPALPHA}"
-
-    def get_compact_qobj(self):
-        return swapalpha(self.arg_value)
-
-
-class MS(TwoQubitGate):
-    def __init__(self, targets, arg_value, **kwargs):
-        super().__init__(targets=targets, arg_value=arg_value, **kwargs)
-        self.name = "MS"
-        self.latex_str = r"{\rm MS}"
-
-    def get_compact_qobj(self):
-        return molmer_sorensen(self.arg_value)
-
-
-class TOFFOLI(Gate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "TOFFOLI"
-        self.latex_str = r"{\rm TOFFOLI}"
-
-    def get_compact_qobj(self):
-        return toffoli()
-
-
-class FREDKIN(Gate):
-    def __init__(self, targets, **kwargs):
-        super().__init__(targets=targets, **kwargs)
-        self.name = "FREDKIN"
-        self.latex_str = r"{\rm FREDKIN}"
-
-    def get_compact_qobj(self):
-        return fredkin()
-
-
-class ControlledGate(Gate):
-    def __init__(
-        self, targets, controls, control_value, target_gate, **kwargs
-    ):
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            control_value=control_value,
-            target_gate=target_gate,
-            **kwargs,
-        )
-        self.controls = (
-            [controls] if not isinstance(controls, list) else controls
-        )
-        self.control_value = control_value
-        self.target_gate = target_gate
-        self.kwargs = kwargs
-        # In the circuit plot, only the target gate is shown.
-        # The control has its own symbol.
-        self.latex_str = target_gate(
-            targets=self.targets, **self.kwargs
-        ).latex_str
-
-    def get_compact_qobj(self):
-        quantum_control_value = (
-            self.control_value[0]
-            if isinstance(self.control_value, Iterable)
-            else self.control_value
-        )
-        return controlled_gate(
-            U=self.target_gate(
-                targets=self.targets, **self.kwargs
-            ).get_compact_qobj(),
-            controls=list(range(len(self.controls))),
-            targets=list(
-                range(
-                    len(self.controls), len(self.targets) + len(self.controls)
-                )
-            ),
-            control_value=quantum_control_value,
-        )
-
-
-class _OneControlledGate(ControlledGate):
-    """
-    This class allows correctly generating the gate instance
-    when a redundant control_value is given, e.g.
-    ``CNOT(1, 0, control_value=1)``,
-    and raise an error if it is 0.
-    """
-
-    def __init__(self, targets, controls, target_gate, **kwargs):
-        _control_value = kwargs.get("control_value", None)
-        if _control_value is not None:
-            if (isinstance(_control_value, int) and _control_value != 1) or (
-                isinstance(_control_value, Iterable) and _control_value[0] != 1
-            ):
-                raise ValueError(
-                    f"{self.__class__.__name__} must has control_value=1"
-                )
-        else:
-            kwargs["control_value"] = 1
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            target_gate=target_gate,
-            **kwargs,
-        )
-
-
-class CNOT(_OneControlledGate):
-    def __init__(self, targets, controls, **kwargs):
-        self.target_gate = X
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            target_gate=self.target_gate,
-            **kwargs,
-        )
-        self.name = "CNOT"
-
-    def get_compact_qobj(self):
-        return cnot()
-
-
-class CSIGN(_OneControlledGate):
-    def __init__(self, targets, controls, **kwargs):
-        self.target_gate = Z
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            target_gate=self.target_gate,
-            **kwargs,
-        )
-        self.name = "CSIGN"
-
-    def get_compact_qobj(self):
-        return csign()
-
-
-class CZ(CSIGN):
-    def __init__(self, targets, controls, **kwargs):
-        self.target_gate = Z
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            **kwargs,
-        )
-        self.name = "CZ"
-
-
-class CPHASE(_OneControlledGate):
-    def __init__(
-        self, targets, controls, arg_value, control_value=1, **kwargs
-    ):
-        self.target_gate = RZ
-        super().__init__(
-            targets=targets,
-            controls=controls,
-            arg_value=arg_value,
-            target_gate=self.target_gate,
-            **kwargs,
-        )
-        self.name = "CPHASE"
-
-    def get_compact_qobj(self):
-        return cphase(self.arg_value)
-
-
-CRY = partial(_OneControlledGate, target_gate=RY)
-CRX = partial(_OneControlledGate, target_gate=RX)
-CRZ = partial(_OneControlledGate, target_gate=RZ)
-CY = partial(_OneControlledGate, target_gate=Y)
-CX = partial(_OneControlledGate, target_gate=X)
-CT = partial(_OneControlledGate, target_gate=T)
-CS = partial(_OneControlledGate, target_gate=S)
-
-GATE_CLASS_MAP = {
-    "X": X,
-    "Y": Y,
-    "Z": Z,
-    "RX": RX,
-    "RY": RY,
-    "RZ": RZ,
-    "H": H,
-    "SNOT": SNOT,
-    "SQRTNOT": SQRTNOT,
-    "S": S,
-    "T": T,
-    "QASMU": QASMU,
-    "SWAP": SWAP,
-    "ISWAP": ISWAP,
-    "iSWAP": ISWAP,
-    "CNOT": CNOT,
-    "SQRTSWAP": SQRTSWAP,
-    "SQRTISWAP": SQRTISWAP,
-    "SWAPALPHA": SWAPALPHA,
-    "SWAPalpha": SWAPALPHA,
-    "MS": MS,
-    "TOFFOLI": TOFFOLI,
-    "FREDKIN": FREDKIN,
-    "CNOT": CNOT,
-    "CSIGN": CSIGN,
-    "CRX": CRX,
-    "CRY": CRY,
-    "CRZ": CRZ,
-    "CY": CY,
-    "CX": CX,
-    "CZ": CZ,
-    "CS": CS,
-    "CT": CT,
-    "CPHASE": CPHASE,
-}
 
 #
 # Single Qubit Gates
